backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
ultralytics
boxmot
opencv-python-headless
torch
torchvision
scikit-learn
ffmpeg-python
scipy
pydantic-settings
numpy
aiofiles
